Bug fixes.
Giuliano:
2) I don't appear in the list of potential supervisors, hence I cannot try the email feature.

3) Can I edit the supervisor email template?

6) Stats page. Applications per months -> Applications per month.

Morris:
In order to see uploaded files  ( application forms + references), you need to click supervisions.   Files & supervisors should by default be shown.

I cannot add myself as supervisor - my own name does not appear in the list.

Edit template for emails is not working - but not that critical as editing emails seem to be working OK and contain the basic information required.